// frontend/ticket-booking-web/src/app/api/model/order-item-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderItemDto { 
    id: string;
    ticketCategoryId: string;
    ticketCategoryName: string;
    quantity: number;
    unitPriceAmount: number;
    currency: string;
    lineTotalAmount: number;
}